#3847, coverity fixes for 2.2
git-svn-id: http://svn.osgeo.org/postgis/branches/2.2@15725 b70326c6-7e19-0410-871a-916f4a2858ee

+++ raster/loader/raster2pgsql.c
@@ -511,7 +511,7 @@ static void
 init_rastinfo(RASTERINFO *info) {
 	info->srid = SRID_UNKNOWN;
 	info->srs = NULL;
-	memset(info->dim, 0, sizeof(double) * 2);
+	memset(info->dim, 0, sizeof(uint32_t) * 2);
 	info->nband_count = 0;
 	info->nband = NULL;
 	info->gdalbandtype = NULL;
